moved the logic and testing for filtering unique and ranking contact mode entries from the Person model to a helper for the view
